Splitting apply.py into lazy/functional and lazy/array
